llm-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
